Changed filter type id to long @rajesh

+++ src/main/java/com/ysyt/dao/IAccomodationDao.java
@@ -68,7 +68,7 @@ public interface IAccomodationDao {
 	Uploads createUpload(Uploads upload, SessionFactory sessionFactory);
 
 
-	Double getMinPriceRange(String minField, List<Long> locationIds, Integer typeId, SessionFactory sessionFactory);
+	Double getMinPriceRange(String minField, List<Long> locationIds, Long typeId, SessionFactory sessionFactory);
 
 	
 }
